# memoize unfiltered sample