The df.sample operation is fairly expensive and occupies a large percentage of the runtime. 
<img src="https://user-images.githubusercontent.com/5554675/94278207-21209600-ff7d-11ea-890c-ff23ec37438e.png" width=300></img>
We memoize any unfiltered samples generated for the dataframe in order to speed up the recommendations. This avoids many repetitive computation to sample for every visualizations. We only operate on unfiltered samples since the filter step occurs before sampling, so the input dataframe would be different. This yields a few seconds of improvement in runtime for generating the Lux display:

- Airbnb [244475 rows x 12 cols] : 19.29s → 18.14 s

- Flights [855632 rows x 11 cols]:  21.24s → 18.80s